requirements.txt
Flask
Flask-SQLAlchemy
gevent
PyMySQL
gunicorn
